refactor: Consolidate test doubles into shared fakes package
- Create FakeUserPreferencesRepository in dev.hossain.mathtutor.fakes package
  - Centralized implementation used across all test files
  - Supports all UserPreferencesRepository methods
  - Includes helper methods for synchronous access in tests

- Remove duplicate FakeUserPreferencesRepository implementations from:
  - DeveloperPortalPresenterTest
  - ParentSettingsPresenterTest
  - AudioHapticSettingsPresenterTest

Benefits:
- Eliminates code duplication in test infrastructure
- Single source of truth for fake implementations
- Easier to maintain and update test doubles
- Follows DRY principle for test code
- All 31 unit tests pass successfully
