+++ app/src/test/java/dev/hossain/mathtutor/ui/settings/AudioHapticSettingsPresenterTest.kt
@@ -4,6 +4,7 @@ import com.google.common.truth.Truth.assertThat
 import dev.hossain.mathtutor.audio.AudioService
 import dev.hossain.mathtutor.data.UserPreferencesRepository
 import dev.hossain.mathtutor.domain.model.GradeLevel
+import dev.hossain.mathtutor.fakes.FakeUserPreferencesRepository
 import dev.hossain.mathtutor.haptic.HapticService
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.MutableStateFlow
@@ -151,142 +152,6 @@ class AudioHapticSettingsPresenterTest {
 
         // Then
         assertThat(hapticService.isHapticsEnabled()).isFalse()
-    }
-
-    /**
-     * Fake implementation of [UserPreferencesRepository] for testing.
-     */
-    private class FakeUserPreferencesRepository : UserPreferencesRepository {
-        private val soundEffectsEnabledFlow = MutableStateFlow(true)
-        private val backgroundMusicEnabledFlow = MutableStateFlow(false)
-        private val hapticsEnabledFlow = MutableStateFlow(true)
-        private val volumeFlow = MutableStateFlow(dev.hossain.mathtutor.audio.AudioConstants.DEFAULT_SOUND_EFFECTS_VOLUME)
-        private val highContrastEnabledFlow = MutableStateFlow(false)
-        private val largeTextEnabledFlow = MutableStateFlow(false)
-        private val analyticsEnabledFlow = MutableStateFlow(true)
-        private val hintSystemEnabledFlow = MutableStateFlow(true)
-
-        override val isOnboardingCompleted: Flow<Boolean> = MutableStateFlow(false)
-
-        override suspend fun setOnboardingCompleted(completed: Boolean) {}
-
-        override val isHapticsEnabled: Flow<Boolean> = hapticsEnabledFlow
-
-        override suspend fun setHapticsEnabled(enabled: Boolean) {
-            hapticsEnabledFlow.value = enabled
-        }
-
-        override val isSoundEffectsEnabled: Flow<Boolean> = soundEffectsEnabledFlow
-
-        override suspend fun setSoundEffectsEnabled(enabled: Boolean) {
-            soundEffectsEnabledFlow.value = enabled
-        }
-
-        override val isBackgroundMusicEnabled: Flow<Boolean> = backgroundMusicEnabledFlow
-
-        override suspend fun setBackgroundMusicEnabled(enabled: Boolean) {
-            backgroundMusicEnabledFlow.value = enabled
-        }
-
-        override val volume: Flow<Float> = volumeFlow
-
-        override suspend fun setVolume(volume: Float) {
-            volumeFlow.value = volume.coerceIn(0f, 1f)
-        }
-
-        override val isHighContrastEnabled: Flow<Boolean> = highContrastEnabledFlow
-
-        override suspend fun setHighContrastEnabled(enabled: Boolean) {
-            highContrastEnabledFlow.value = enabled
-        }
-
-        override val isLargeTextEnabled: Flow<Boolean> = largeTextEnabledFlow
-
-        override suspend fun setLargeTextEnabled(enabled: Boolean) {
-            largeTextEnabledFlow.value = enabled
-        }
-
-        override val isAnalyticsEnabled: Flow<Boolean> = analyticsEnabledFlow
-
-        override suspend fun setAnalyticsEnabled(enabled: Boolean) {
-            analyticsEnabledFlow.value = enabled
-        }
-
-        private val importGuideExpandedFlow = MutableStateFlow(true)
-
-        override val isImportGuideExpanded: Flow<Boolean> = importGuideExpandedFlow
-
-        override suspend fun setImportGuideExpanded(expanded: Boolean) {
-            importGuideExpandedFlow.value = expanded
-        }
-
-        private val gameTrialAttemptsFlows = mutableMapOf<dev.hossain.mathtutor.domain.model.Game, MutableStateFlow<Int>>()
-
-        override fun getGameTrialAttempts(game: dev.hossain.mathtutor.domain.model.Game): Flow<Int> =
-            gameTrialAttemptsFlows.getOrPut(game) { MutableStateFlow(0) }
-
-        override suspend fun incrementGameTrialAttempts(game: dev.hossain.mathtutor.domain.model.Game): Int {
-            val flow = gameTrialAttemptsFlows.getOrPut(game) { MutableStateFlow(0) }
-            val newCount = (flow.value + 1).coerceAtMost(3)
-            flow.value = newCount
-            return newCount
-        }
-
-        // Test helper methods
-        fun setSoundEffectsEnabledSync(enabled: Boolean) {
-            soundEffectsEnabledFlow.value = enabled
-        }
-
-        fun getCurrentSoundEffectsEnabled(): Boolean = soundEffectsEnabledFlow.value
-
-        fun setBackgroundMusicEnabledSync(enabled: Boolean) {
-            backgroundMusicEnabledFlow.value = enabled
-        }
-
-        fun getCurrentBackgroundMusicEnabled(): Boolean = backgroundMusicEnabledFlow.value
-
-        fun setHapticsEnabledSync(enabled: Boolean) {
-            hapticsEnabledFlow.value = enabled
-        }
-
-        fun getCurrentHapticsEnabled(): Boolean = hapticsEnabledFlow.value
-
-        fun setVolumeSync(volume: Float) {
-            volumeFlow.value = volume.coerceIn(0f, 1f)
-        }
-
-        fun getCurrentVolume(): Float = volumeFlow.value
-
-        fun setHighContrastEnabledSync(enabled: Boolean) {
-            highContrastEnabledFlow.value = enabled
-        }
-
-        fun getCurrentHighContrastEnabled(): Boolean = highContrastEnabledFlow.value
-
-        fun setLargeTextEnabledSync(enabled: Boolean) {
-            largeTextEnabledFlow.value = enabled
-        }
-
-        fun getCurrentLargeTextEnabled(): Boolean = largeTextEnabledFlow.value
-
-        // Parent control methods (minimal implementation for testing)
-        override val parentPinHash: Flow<String?> = MutableStateFlow(null)
-
-        override suspend fun setParentPin(pin: String) {}
-
-        override suspend fun verifyParentPin(pin: String): Boolean = false
-
-        override suspend fun clearParentPin() {}
-
-        override val maxGradeLevel: Flow<GradeLevel?> = MutableStateFlow(null)
-
-        override suspend fun setMaxGradeLevel(gradeLevel: GradeLevel?) {}
-
-        override val isHintSystemEnabled: Flow<Boolean> = hintSystemEnabledFlow
-
-        override suspend fun setHintSystemEnabled(enabled: Boolean) {
-            hintSystemEnabledFlow.value = enabled
-        }
     }
 
     /**
